Add MetaData
Makes the add-on installable. 
https://wiki.blender.org/wiki/Process/Addons/Guidelines/metainfo

You properly know about this, but for newcomers, this may be the difference between "working" and not "working". I hope that's okay?

Add-on works fine now. Getting the source dimensions shouldn't be this complicated, maybe file a bug here: https://developer.blender.org/project/view/28/

Actually, there is a request at right-click select for this fix/add-on: https://blender.community/c/rightclickselect/GXbbbc/vse-aspect-correction